fix(flash-moe): address round-4 review — restore worker flash reads + legacy shims
distributed_worker.py reads promoted Flash fields from settings.* (the
rebase squash incorrectly routed them back through experimental.*, which
would AttributeError at startup since #274/#327 moved them to Settings).

worker_main restores the surface_legacy_flash_env() call (dropped by the
rebase) so direct ``python -m olmlx.engine.distributed_worker`` invocations
keep the one-release deprecation window for OLMLX_EXPERIMENTAL_FLASH*.

Move surface_legacy_flash_moe_env (and its dotenv/forwarder helpers) from
olmlx.cli to olmlx.config so the worker can pair it with the existing
flash shim without pulling in argparse/uvicorn. The .env quote check now
requires *paired* quote characters before suppressing inline-comment
stripping, fixing the mixed-quote edge case.

ExpertCacheStats.load_failures now drains all futures before re-raising,
so failures is a true per-expert count (consistent with cache_hits and
cache_misses) instead of capped at 1 per call.

Also fixes Ruff F841/F401 failures (unused resolved_exp, unused
experimental import in cli.py, unused registry import in test_registry.py).
